Converted Knowledge Graph with SSO for our nodes
Each person's URI is now built off of the base SSO URI which exists at a
currently fictional "http://david.jefts/sso". This sets the base for the
Knowledge Graph and the hub for all of the external ontologies that will
be connected now and in the future.

The Facebook URLs are now used as the URIs for foaf:OnlineAccount, and
the graph is created with that assumption.

Knowledge represented in the Graph includes the logged-in user (testing
with John Keck), their friends, and all of their OnlineAccounts.